Calculate plan in f64 almost everywhere.

+++ src/earth/coordinate.rs
@@ -1,5 +1,6 @@
-use crate::util::lat_long_format::LatLongFormat;
 use std::f64::consts::PI;
+
+use crate::util::lat_long_format::LatLongFormat;
 
 #[derive(Debug, Clone, PartialEq)]
 pub struct Coordinate {
@@ -43,7 +44,7 @@ impl Coordinate {
         self.bearing_to(l).to_degrees()
     }
 
-    pub fn coordinate_at(&self, distance: i32, heading: f64) -> Coordinate {
+    pub fn coordinate_at(&self, distance: f64, heading: f64) -> Coordinate {
         let d = distance as f64 / Self::EARTH_RADIUS;
         let lat1 = self.latitude.to_radians();
         let lon1 = self.longitude.to_radians();
@@ -56,7 +57,7 @@ impl Coordinate {
         Coordinate::new(lat.to_degrees(), lon.to_degrees())
     }
 
-    pub fn distance_to(&self, l: &Coordinate) -> i32 {
+    pub fn distance_to(&self, l: &Coordinate) -> f64 {
         let lat1 = self.latitude.to_radians();
         let lat2 = l.latitude.to_radians();
         let lon1 = self.longitude.to_radians();
@@ -67,7 +68,7 @@ impl Coordinate {
         let a = (d_lat / 2.0).sin().powi(2) + lat1.cos() * lat2.cos() * (d_lon / 2.0).sin().powi(2);
         let d = 2.0 * a.sqrt().atan2((1.0 - a).sqrt());
 
-        (Self::EARTH_RADIUS * d.abs()) as i32
+        (Self::EARTH_RADIUS * d.abs())
     }
 
     pub fn get_latitude(&self) -> f64 {
@@ -104,22 +105,22 @@ mod tests {
     fn test_distance_to() {
         let c1 = Coordinate::new(-34.0, 151.0);
         let c2 = Coordinate::new(-34.0, 151.0);
-        assert_eq!(c1.distance_to(&c2), 0);
+        assert_eq!(c1.distance_to(&c2), 0.0);
         let c1 = Coordinate::new(-34.0, 151.0);
         let c2 = Coordinate::new(-34.0, 150.0);
-        assert_eq!(c1.distance_to(&c2), 50);
+        assert_eq!(c1.distance_to(&c2), 50.0);
         let c1 = Coordinate::new(-34.0, 151.0);
         let c2 = Coordinate::new(-35.0, 151.0);
-        assert_eq!(c1.distance_to(&c2), 60);
+        assert_eq!(c1.distance_to(&c2), 60.0);
         let c1 = Coordinate::new(-34.45, 150.50);
         let c2 = Coordinate::new(-34.18, 150.86);
-        assert_eq!(c1.distance_to(&c2), 24);
+        assert_eq!(c1.distance_to(&c2), 24.0);
         let c1 = Coordinate::new(-34.0, 151.0);
         let c2 = Coordinate::new(35.0, -151.0);
-        assert_eq!(c1.distance_to(&c2), 5272);
+        assert_eq!(c1.distance_to(&c2), 5272.0);
         let c1 = Coordinate::new(-34.0, 151.0);
         let c2 = Coordinate::new(0.0, 0.0);
-        assert_eq!(c1.distance_to(&c2), 8198);
+        assert_eq!(c1.distance_to(&c2), 8198.0);
     }
 
     #[test]
@@ -138,17 +139,17 @@ mod tests {
     #[test]
     fn test_coordinate_at() {
         let c1 = Coordinate::new(0.0, 151.0);
-        let c2 = c1.coordinate_at(120, 60.0);
+        let c2 = c1.coordinate_at(120.0, 60.0);
         assert!(is_between(c2.latitude, 0.99, 1.01));
         assert!(is_between(c2.longitude, 152.72, 152.74));
 
         let c1 = Coordinate::new(-34.0, 151.0);
-        let c2 = c1.coordinate_at(120, 120.0);
+        let c2 = c1.coordinate_at(120.0, 120.0);
         assert!(is_between(c2.latitude, -34.99, -34.97));
         assert!(is_between(c2.longitude, 153.10, 153.12));
 
         let c1 = Coordinate::new(-34.0, 151.0);
-        let c2 = c1.coordinate_at(100000, 120.0);
+        let c2 = c1.coordinate_at(100000.0, 120.0);
         assert!(is_between(c2.latitude, 43.0, 44.0));
         assert!(is_between(c2.longitude, 28.0, 29.0));
     }
